# backend/app/schemas/auth.py
import uuid

from pydantic import BaseModel, EmailStr


class SignupRequest(BaseModel):
    email: EmailStr
    password: str
    display_name: str
    profile: "BackgroundProfileUpdate"


class SigninRequest(BaseModel):
    email: EmailStr
    password: str


class UserResponse(BaseModel):
    id: uuid.UUID
    email: str
    display_name: str

    model_config = {"from_attributes": True}


class AuthResponse(BaseModel):
    user: UserResponse
    message: str


from app.schemas.profile import BackgroundProfileUpdate  # noqa: E402

SignupRequest.model_rebuild()